test(forms): Add utility functions and update test files to use them
Moves common helper functions into a shared test utility to reduce duplication.
