+++ packages/forms/signals/test/node/form_root.spec.ts
@@ -11,6 +11,7 @@ import {TestBed} from '@angular/core/testing';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 
 import {form, FormRoot} from '../../public_api';
+import {act} from '@angular/private/testing';
 
 @Component({
   template: `
@@ -143,11 +144,3 @@ describe('FormRoot', () => {
     expect(component.submitted).toBeTrue();
   });
 });
-
-function act<T>(fn: () => T): T {
-  try {
-    return fn();
-  } finally {
-    TestBed.tick();
-  }
-}
