+++ packages/forms/signals/test/web/compat_form.spec.ts
@@ -11,6 +11,7 @@ import {TestBed} from '@angular/core/testing';
 import {FormControl} from '@angular/forms';
 import {compatForm} from '../../compat';
 import {FormField} from '../../public_api';
+import {act} from '@angular/private/testing';
 
 describe('compatForm with [formField] directive', () => {
   beforeEach(() => {
@@ -87,11 +88,3 @@ describe('compatForm with [formField] directive', () => {
     expect(fixture.componentInstance.f().value().species).toBe('cat');
   });
 });
-
-function act<T>(fn: () => T): T {
-  try {
-    return fn();
-  } finally {
-    TestBed.tick();
-  }
-}
