+++ packages/forms/signals/test/web/focus.spec.ts
@@ -6,20 +6,12 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {
-  ApplicationRef,
-  Component,
-  inject,
-  input,
-  model,
-  signal,
-  viewChild,
-  type ElementRef,
-} from '@angular/core';
+import {Component, inject, input, model, signal, viewChild, type ElementRef} from '@angular/core';
 import {TestBed} from '@angular/core/testing';
 import {FormControl} from '@angular/forms';
 import {compatForm} from '../../compat';
 import {FormField, form, type Field, type FieldTree} from '../../public_api';
+import {act} from '@angular/private/testing';
 
 describe('FieldState focus behavior', () => {
   it('should focus a native control', async () => {
@@ -31,11 +23,11 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const input = fixture.nativeElement.firstChild;
 
     expect(document.activeElement).not.toBe(input);
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(input);
   });
 
@@ -64,10 +56,10 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const customControl = fixture.nativeElement.firstChild as HTMLInputElement;
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(focusCalled).toBeTrue();
     expect(document.activeElement).not.toBe(customControl);
     expect(document.activeElement).toBe(customControl.querySelector('input'));
@@ -125,10 +117,10 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const customControl = fixture.nativeElement.firstChild as HTMLInputElement;
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(customControl);
   });
 
@@ -147,16 +139,16 @@ describe('FieldState focus behavior', () => {
       showFirst = signal(false);
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const input2 = fixture.nativeElement.querySelector('#input2');
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(input2);
 
-    await act(() => fixture.componentInstance.showFirst.set(true));
+    act(() => fixture.componentInstance.showFirst.set(true));
     const input1 = fixture.nativeElement.querySelector('#input1');
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(input1);
   });
 
@@ -181,10 +173,10 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal({child1: '', child2: ''}));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const child2 = fixture.nativeElement.querySelector('#child2');
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(child2);
   });
 
@@ -197,11 +189,11 @@ describe('FieldState focus behavior', () => {
       readonly f = compatForm(signal(new FormControl('', {nonNullable: true})));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const input = fixture.nativeElement.firstChild;
 
     expect(document.activeElement).not.toBe(input);
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(input);
   });
 
@@ -222,10 +214,10 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
 
     const focusedEl = document.activeElement;
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(focusedEl);
   });
 
@@ -253,11 +245,11 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const nativeInput = fixture.nativeElement.querySelector('custom-control > input');
     expect(nativeInput).toBeTruthy();
 
-    await act(() => fixture.componentInstance.f().focusBoundControl());
+    act(() => fixture.componentInstance.f().focusBoundControl());
     expect(document.activeElement).toBe(nativeInput);
   });
 
@@ -270,12 +262,12 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
     const input = fixture.nativeElement.firstChild as HTMLInputElement;
 
     const focusSpy = spyOn(input, 'focus');
 
-    await act(() => fixture.componentInstance.f().focusBoundControl({preventScroll: true}));
+    act(() => fixture.componentInstance.f().focusBoundControl({preventScroll: true}));
     expect(focusSpy).toHaveBeenCalledWith({preventScroll: true});
   });
 
@@ -302,15 +294,9 @@ describe('FieldState focus behavior', () => {
       readonly f = form(signal(''));
     }
 
-    const fixture = await act(() => TestBed.createComponent(TestCmp));
+    const fixture = act(() => TestBed.createComponent(TestCmp));
 
-    await act(() => fixture.componentInstance.f().focusBoundControl({preventScroll: true}));
+    act(() => fixture.componentInstance.f().focusBoundControl({preventScroll: true}));
     expect(receivedOptions).toEqual({preventScroll: true});
   });
 });
-
-async function act<T>(fn: () => T): Promise<T> {
-  const result = fn();
-  await TestBed.inject(ApplicationRef).whenStable();
-  return result;
-}
