+++ packages/forms/signals/test/web/signal_form_control_web.spec.ts
@@ -9,10 +9,10 @@
 import {Component, Injector, inject, provideZonelessChangeDetection, signal} from '@angular/core';
 import {TestBed} from '@angular/core/testing';
 import {FormControl, FormGroup, ReactiveFormsModule} from '@angular/forms';
-import {disabled} from '@angular/forms/signals';
+import {disabled, FormField} from '@angular/forms/signals';
 
 import {SignalFormControl} from '../../compat';
-import {FormField} from '../../src/directive/form_field';
+import {act} from '@angular/private/testing';
 
 describe('SignalFormControl (web)', () => {
   beforeEach(() => {
@@ -24,7 +24,6 @@ describe('SignalFormControl (web)', () => {
 
   it('binds to formField directive', () => {
     @Component({
-      standalone: true,
       imports: [ReactiveFormsModule, FormField],
       template: `<input [formField]="signalControl.fieldTree" />`,
     })
@@ -51,7 +50,6 @@ describe('SignalFormControl (web)', () => {
 
   it('binds inside nested FormGroup via formGroupName', () => {
     @Component({
-      standalone: true,
       imports: [ReactiveFormsModule, FormField],
       template: `
         <div [formGroup]="group">
@@ -90,7 +88,6 @@ describe('SignalFormControl (web)', () => {
 
   it('should unregister disabled callback when directive is destroyed', () => {
     @Component({
-      standalone: true,
       imports: [ReactiveFormsModule],
       template: `
         @if (showInput()) {
@@ -123,11 +120,3 @@ describe('SignalFormControl (web)', () => {
     }).not.toThrow();
   });
 });
-
-function act<T>(fn: () => T): T {
-  try {
-    return fn();
-  } finally {
-    TestBed.tick();
-  }
-}
